Starting with the docs
I'm applying something that I don't even know if it's a thing, but I
call DTDD: doctest driven development. The good thing about it is that
you end up with a documented and unit tested API, and it forces me to
think of ways to implement it using pure functions, what makes it easy
to test further scenarios.